requirements.txt
anthropic>=0.20.0
streamlit>=1.32.0
langchain>=0.1.0
langchain-community>=0.0.20
chromadb>=0.4.22
sentence-transformers>=2.5.1
pypdf>=4.0.0
python-dotenv>=1.0.0
